requirements.txt
wandb
scikit-spatial
mujoco
rerun-sdk
torch
pytorch-lightning
seaborn